Move entity collect and search relations code to DocumentLoader, so it would load for every way of map loading
